% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categ}
\alias{categ}
\title{Table used to convert unscaled values to categories}
\format{A data frame with 53 rows and 6 variables:
\describe{
  \item{indicateur}{code of the indicator}
  \item{nom_indicateur}{full name of the indicator}
  \item{TD}{threshold for très défavorable}
  \item{D}{threshold for défavorable}
  \item{I}{threshold for intermédiaire}
  \item{F}{threshold for favorable}
}}
\usage{
categ
}
\description{
This table lists, for each indicator, the thresholds used to categorise them.
}
\keyword{datasets}
