% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_diag.R
\name{compile_diag}
\alias{compile_diag}
\title{Compilation de diagnostics IDEA4}
\usage{
compile_diag(input_dir, perfecto_path)
}
\arguments{
\item{input_dir}{chemin vers le dossier avec des calculateurs}

\item{perfecto_path}{chemin vers la base perfecto}
}
\value{
exporte une base de donnees dans le dossier "BDD"
}
\description{
Compilation de diagnostics IDEA4
}
