% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decision_rules_total}
\alias{decision_rules_total}
\title{List used to convert indicators to properties}
\format{A list of length 47}
\usage{
decision_rules_total
}
\description{
This list contains, for each node, the decision table created by the authors
}
\keyword{datasets}
