% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagIDEA.R
\name{DiagIDEA}
\alias{DiagIDEA}
\title{Wrapper function for IDEATools}
\usage{
DiagIDEA(
  input,
  output = paste0("RES_", Sys.Date()),
  dimensions = TRUE,
  polar = TRUE,
  trees = TRUE,
  radar = TRUE,
  meta = TRUE,
  export = FALSE
)
}
\arguments{
\item{input}{a system path leading either to a single file or a directory. If the input is a single file, accepted formats are : .xls, .xlsx and .json.}

\item{output}{the output directory}

\item{dimensions}{Should the diagnosis feature the dimension plots ?}

\item{trees}{Should the diagnosis feature the enlightened trees ?}

\item{radar}{Should the diagnosis feature the radar plots ?}

\item{meta}{Should the diagnosis feature the meta-analysis plots ?}

\item{export}{Should the result be exported as png/pdf files, or as an R object ?}
}
\value{
a named list, containing :
\describe{
  \item{dimensions}{a string which can be "single" or "folder" depending on the input type.}
  \item{trees}{a tibble dataframe with the extracted data}
  \item{radars}{results of the property analysis}
  \item{meta}{metadata extracted from the input}
}
}
\description{
Wrapper function for IDEATools
}
\examples{
library(IDEATools)
path <- system.file("example_json.json", package = "IDEATools")
Diag <- DiagIDEA(path,"tmp", meta = FALSE, export = FALSE)
}
