% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionsPlots.R
\name{dimensionsPlots}
\alias{dimensionsPlots}
\title{Generate adequate plots for the dimension approach of the IDEA method}
\usage{
dimensionsPlots(IDEAdata)
}
\arguments{
\item{IDEAdata}{output generated by an `importIDEA()` call}
}
\value{
a named list of plots
}
\description{
Generate adequate plots for the dimension approach of the IDEA method
}
\examples{
library(IDEATools)
path <- system.file("example.xls", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
IDEAres <- dimensionsPlots(IDEAdata)
}
