% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{find_pos}
\alias{find_pos}
\title{Finds the position of the given rectangle in the svg string}
\usage{
find_pos(start, end, choice, car)
}
\arguments{
\item{start}{the starting position of the rectangle}

\item{end}{the ending position of the rectangle}

\item{choice}{output type, can be "id" or "style"}

\item{car}{the svg string canvas}
}
\value{
the correct rectangle id or style position
}
\description{
Finds the position of the given rectangle in the svg string
}
