% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIDEA2.R
\name{importIDEA2}
\alias{importIDEA2}
\title{Import IDEA4 data}
\usage{
importIDEA2(input, anonymous = FALSE)
}
\arguments{
\item{input}{a system path leading either to a single file or a directory. If the input is a single file, accepted formats are : .xls, .xlsx and .json.}

\item{anonymous}{Boolean. Should the results be anonymised ?}
}
\value{
a named list, containing :
    analysis.type : a string which can be "single" or "folder" depending on the input type.
    dataset : a tibble dataframe with the extracted data
    nodes : results of the property analysis
    metadata : metadata extracted from the input
}
\description{
Import IDEA4 data
}
\examples{
library(IDEATools)
path <- system.file("example.xls", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
}
