% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jsonify2}
\alias{jsonify2}
\title{Convert IDEA excel files to json, without output}
\usage{
jsonify2(input, output)
}
\arguments{
\item{input}{the directory containing the xls/xlsx files to convert}

\item{output}{the directory where the json files will be created}
}
\value{
a list of json files exported to the output directory
}
\description{
Convert IDEA excel files to json, without output
}
