% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label_nodes}
\alias{label_nodes}
\title{Table gathering all metadata on IDEA variables}
\format{A data frame with 99 rows and 7 variables:
\describe{
  \item{code_indicateur}{code of the indicator}
  \item{nom_indicateur}{full name of the indicator}
  \item{nom_complet}{the code and full name of the indicator collapsed together}
  \item{level}{The levels of the given indicator in the heuristic map}
  \item{dimension}{Dimension of the indicator}
  \item{composante}{compound of the indicator}
}}
\usage{
label_nodes
}
\description{
This table contains info for each IDEA variable about the full label, heuristic map level, dimension, etc.
}
\keyword{datasets}
