% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeExcel.R
\name{MakeExcel}
\alias{MakeExcel}
\title{Render an Excel spreadsheet of an IDEA diagnosis}
\usage{
MakeExcel(input, outdir = getwd(), silent = FALSE, append = TRUE)
}
\arguments{
\item{outdir}{the output directory name. Defaults to your working directory.}

\item{silent}{Should the algorithm be silent ?}

\item{append}{Should the results be appended to the original data ? This only works with the `.xlsx` extension}
}
\value{
An excel spreadsheet
}
\description{
Render an Excel spreadsheet of an IDEA diagnosis
}
\examples{
library(IDEATools)
path <- system.file("example_json.json", package = "IDEATools")
MakeExcel(IDEAdata)
}
