% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGroupExcel.R
\name{MakeGroupExcel}
\alias{MakeGroupExcel}
\title{Render an Excel spreadsheet of an IDEA diagnosis}
\usage{
MakeGroupExcel(
  input,
  output_dir = getwd(),
  file = "analyse_de_groupe",
  silent = FALSE
)
}
\arguments{
\item{input}{a system path leading either to a single file or a directory. If the input is a single file, accepted formats are : .xls, .xlsx and .json.}

\item{output_dir}{the output directory name. Defaults to your working directory.}

\item{silent}{Should the algorithm be silent ?}

\item{append}{Should the results be appended to the original data ? This only works with the `.xlsx` extension}
}
\value{
An excel spreadsheet
}
\description{
Render an Excel spreadsheet of an IDEA diagnosis
}
\examples{
library(IDEATools)
path <- system.file("", package = "IDEATools")
MakeGroupExcel(path)
}
