% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGroupReport.R
\name{MakeGroupReport}
\alias{MakeGroupReport}
\title{Render a PDF report of an IDEA diagnosis}
\usage{
MakeGroupReport(
  input,
  output_dir = getwd(),
  file = "Rapport_groupe",
  silent = FALSE
)
}
\arguments{
\item{file}{the file name with no extension}

\item{silent}{Should the algorithm be silent ?}

\item{IDEAdata}{output generated by an `importIDEA()` call}
}
\value{
A PDF report
}
\description{
Render a PDF report of an IDEA diagnosis
}
\examples{
library(IDEATools)
path <- system.file("", package = "IDEATools")
MakeGroupReport(path)
}
