% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePagedownGroupReport.R
\name{MakePagedownGroupReport}
\alias{MakePagedownGroupReport}
\title{Render a HTML/PDF report of an IDEA diagnosis}
\usage{
MakePagedownGroupReport(
  input,
  output_dir = getwd(),
  file = "Rapport_groupe",
  silent = FALSE
)
}
\arguments{
\item{input}{a system path leading either to a single file or a directory. If the input is a single file, accepted formats are : .xls, .xlsx and .json.}

\item{output_dir}{the output directory}

\item{silent}{Should the algorithm be silent ?}
}
\value{
A ODT report
}
\description{
Render a HTML/PDF report of an IDEA diagnosis
}
\examples{
library(IDEATools)
path <- system.file("", package = "IDEATools")
MakePagedownGroupReport(path)
}
