% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeReport.R
\name{MakeReport}
\alias{MakeReport}
\title{Render a PDF report of an IDEA diagnosis}
\usage{
MakeReport(
  IDEAdata,
  file = paste0("Rapport_", Sys.Date(), ".pdf"),
  silent = FALSE
)
}
\arguments{
\item{IDEAdata}{output generated by an `importIDEA()` call}

\item{file}{the output file name. Should end by ".pdf"}

\item{silent}{Should the algorithm be silent ?}
}
\value{
A PDF report
}
\description{
Render a PDF report of an IDEA diagnosis
}
\examples{
library(IDEATools)
path <- system.file("example_json.json", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
MakeReport(IDEAdata)
}
