% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrees.R
\name{MakeTrees}
\alias{MakeTrees}
\title{Generate colored heuristic maps for the IDEA method}
\usage{
MakeTrees(IDEAdata)
}
\arguments{
\item{IDEAdata}{output generated by an `importIDEA()` call}
}
\value{
a named list of plots
}
\description{
Generate colored heuristic maps for the IDEA method
}
\examples{
library(IDEATools)
path <- system.file("example_json.json", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
IDEAres <- MakeTrees(IDEAdata)
}
