% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeWordGroupReport.R
\name{MakeWordGroupReport}
\alias{MakeWordGroupReport}
\title{Render a Word report of an IDEA diagnosis}
\usage{
MakeWordGroupReport(
  input,
  output_dir = getwd(),
  file = "Rapport_groupe",
  silent = FALSE
)
}
\arguments{
\item{input}{a system path leading either to a single file or a directory. If the input is a single file, accepted formats are : .xls, .xlsx and .json.}

\item{output_dir}{the output directory}

\item{silent}{Should the algorithm be silent ?}
}
\value{
A PDF report
}
\description{
Render a Word report of an IDEA diagnosis
}
\examples{
library(IDEATools)
path <- system.file("", package = "IDEATools")
MakeWordGroupReport(path)
}
