% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaIDEA.R
\name{metaIDEA}
\alias{metaIDEA}
\title{Generate adequate plots for group analysis with the IDEA method}
\usage{
metaIDEA(IDEAdata)
}
\arguments{
\item{IDEAdata}{output generated by an `importIDEA()` call}
}
\value{
a named list of plots
}
\description{
Generate adequate plots for group analysis with the IDEA method
}
\examples{
library(IDEATools)
path <- system.file("", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
IDEAres <- metaIDEA(IDEAdata)
}
