% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolarComponent.R
\name{PolarComponent}
\alias{PolarComponent}
\title{Generate polarised plots for components}
\usage{
PolarComponent(IDEAdata)
}
\arguments{
\item{IDEAdata}{output generated by an `importIDEA()` call}
}
\value{
a named list of plots
}
\description{
Generate polarised plots for components
}
\examples{
library(IDEATools)
path <- system.file("example_json.json", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
IDEAres <- PolarComponent(IDEAdata)
}
