% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryWebIDEA.R
\name{QueryWebIDEA}
\alias{QueryWebIDEA}
\title{Query the WEBIDEA database}
\usage{
QueryWebIDEA(user = "postgres", password = "", id_organisme = 1,
  id_dossier = "654321", annee = 2009)
}
\arguments{
\item{id_organisme}{identifiant de l'organisme}

\item{id_dossier}{identifiant du dossier}

\item{anneee}{year}
}
\value{
a named list, containing :
    analysis.type : a string which can be "single" or "folder" depending on the input type.
    dataset : a tibble dataframe with the extracted data
    nodes : results of the property analysis
    metadata : metadata extracted from the input
}
\description{
Query the WEBIDEA database
}
\examples{
library(IDEATools)
path <- system.file("example_json.json", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
}
