% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarPlots.R
\name{radarPlots}
\alias{radarPlots}
\title{Generate adequate plots for the property approach of the IDEA method}
\usage{
radarPlots(IDEAdata)
}
\arguments{
\item{IDEAdata}{output generated by an `importIDEA()` call}
}
\value{
a named list of plots
}
\description{
Generate adequate plots for the property approach of the IDEA method
}
\examples{
library(IDEATools)
path <- system.file("example_json.json", package = "IDEATools")
IDEAdata <- importIDEA(path, anonymous = FALSE)
IDEAres <- radarPlots(IDEAdata)
}
