% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{replace_col}
\alias{replace_col}
\title{Atribute valid color to a property}
\usage{
replace_col(resultat)
}
\arguments{
\item{resultat}{the value affected to the property}
}
\value{
the correct R color value
}
\description{
Atribute valid color to a property
}
