% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{replace_indicateur}
\alias{replace_indicateur}
\title{Attribute a code to each node}
\usage{
replace_indicateur(indicateur)
}
\arguments{
\item{indicateur}{the name of the node}
}
\value{
a string. the new code for this node
}
\description{
Attribute a code to each node
}
