% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runGUI}
\alias{runGUI}
\title{Minimal Shiny app for IDEATools}
\usage{
runGUI()
}
\description{
This function loads a graphical user interface (GUI) to use IDEATools.
}
\examples{
library(IDEATools)
\dontrun{
runGUI()
}
}
