% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rules.R
\name{show_canvas}
\alias{show_canvas}
\title{Show the blank tree canvas}
\usage{
show_canvas(directory = getwd())
}
\arguments{
\item{directory}{the directory where to output the blank canvas}
}
\description{
Show the blank tree canvas
}
