% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rules.R
\name{show_decision_rules}
\alias{show_decision_rules}
\title{Show decision rules}
\usage{
show_decision_rules(directory = getwd())
}
\arguments{
\item{directory}{the directory where to output the decision rules}
}
\description{
Show decision rules
}
