% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rules.R
\name{show_tree_structure}
\alias{show_tree_structure}
\title{Show the reference table used for building colored trees}
\usage{
show_tree_structure(directory)
}
\arguments{
\item{directory}{the directory where to output the reference tables}
}
\value{
Exports an excel file in the desired directory
}
\description{
Show the reference table used for building colored trees
}
\examples{
library(IDEATools)
show_tree_structure(tempdir())

}
