% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{simplify_indicator_name}
\alias{simplify_indicator_name}
\title{Simplify indicator name if separated by a space}
\usage{
simplify_indicator_name(name)
}
\arguments{
\item{name}{The indicator name}
}
\value{
a string. The simplified indicator name.
}
\description{
Simplify indicator name if separated by a space
}
