% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_idea}
\alias{theme_idea}
\title{IDEA ggplot2 theme}
\usage{
theme_idea(base_size = 15, base_family = "")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base family}
}
\value{
a ggplot2 theme for IDEA
}
\description{
IDEA ggplot2 theme
}
