% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{wrapit}
\alias{wrapit}
\title{Adds a line break for too long strings}
\usage{
wrapit(text)
}
\arguments{
\item{text}{the string to modify}
}
\value{
a the string with line breaks
}
\description{
Adds a line break for too long strings
}
