% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionsPlots.R
\name{dimensionsPlots}
\alias{dimensionsPlots}
\title{Plot visual outputs linked to the "Dimension" approach for IDEA4}
\usage{
dimensionsPlots(IDEAdata)
}
\arguments{
\item{IDEAdata}{output of the `importIDEA()` function}

\item{output_dir}{If a directory path is given, each plot will be exported as a png file in this directory. If the directory doesn't exist yet, it will be created.}
}
\value{
a named list of plots
}
\description{
Plot visual outputs linked to the "Dimension" approach for IDEA4
}
\examples{
path <- system.file("example.xls", package = "RIDEATools")
dat <- importIDEA(path, anonymous = FALSE)
plots <- dimensionsPlots(dat, output_dir = NULL)
}
