% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrees.R
\name{MakeTrees}
\alias{MakeTrees}
\title{Plot visual outputs linked to the "Property" approach for IDEA4}
\usage{
MakeTrees(IDEAdata)
}
\arguments{
\item{IDEAdata}{output of the `importIDEA()` function}
}
\value{
a named list of plots
}
\description{
Plot visual outputs linked to the "Property" approach for IDEA4
}
\examples{
path <- system.file("example.xls", package = "RIDEATools")
dat <- importIDEA(path, anonymous = FALSE)
plots <- dimensionsPlots(dat, output_dir = NULL)
}
