% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rules.R
\name{show_decision_rules}
\alias{show_decision_rules}
\title{Show decision rules}
\usage{
show_decision_rules(directory)
}
\arguments{
\item{directory}{the directory where to output the decision rules}
}
\value{
Exports an excel file in the desired directory
}
\description{
Show decision rules
}
\examples{
library(IDEATools)
show_decision_rules(tempdir())

}
